pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "kaggle>=1.7.4.5",
    "keras>=2.14.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "openai>=1.86.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "scikit-learn>=1.7.0",
    "streamlit>=1.45.1",
    "tensorflow>=2.14.0",
]
